pyproject.toml
[project]
name = "pydantic-ai-chat-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Sam Pratt", email = "samj.pratt@gmail.com" }]
requires-python = ">=3.13"
dependencies = ["pydantic-ai-slim>=1.0"]

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
  "ruff>=0.13.1",
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5",
]

[tool.ruff]
target-version = "py313"
line-length = 88
indent-width = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
